[rename] SeatLayoutTab -> SeatInfoTab 컴포넌트명 변경
해당 컴포넌트는 '좌석 정보' 탭 클릭했을 때 하단에 뜨는 컴포넌트.
레이아웃 정보 뿐만 아니라 다른 정보도 포함하고 있으므로 Layout 대신 Info로 네이밍 변경함.
