+++ src/components/store/SeatInfoTab/SeatInfoTab.styled.ts
