--- src/components/store/SeatLayoutTab/index.ts
@@ -1,1 +1,0 @@
-export * from './SeatLayoutTab';
